pyproject.toml
[tool.poetry]
name = "pylutron-leap"
version = "0.1.0"
description = ""
authors = ["Derek Ditch <dcode@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
marshmallow-dataclass = {extras = ["enum", "union"], version = "^8.5.3"}
marshmallow-enum = "^1.5.1"
marshmallow-union = "^0.1.15"
aioopenssl = "^0.6.0"

[tool.poetry.dev-dependencies]
pytest = "^6.0"
black = "^22.1.0"
flake8 = "^4.0.1"
pylint = "^2.12.2"
mypy = "^0.931"
pydocstyle = "^6.1.1"
pytest-cov = "^3.0.0"
ipykernel = {version = "^6.9.1", allow-prereleases = true}
pytest-flake8 = "^1.0"
pytest-black = "^0.3.12"
pytest-isort = "^3.0.0"
pytest-datadir = "^1.3.1"
python-dotenv = "^0.20.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--black --isort --flake8"
testpaths = ["."]
flake8-max-line-length = 88
flake8-ignore = "E203 E266 E501 W503"

[tool.isort]
profile = "black"

[tool.pyright]
include = ["pylutron_leap", "tests"]
exclude = ["**/__pycache__"]